use logos instead of regex to fix a lot of issues

+++ src/parse.rs
@@ -1,12 +1,10 @@
-use anyhow::{bail, Context as _, IntoResult, Result};
-use chrono::Utc;
-use chrono_tz::Tz;
-use regex::{Captures, Regex, RegexBuilder};
+use std::fmt::Write;
+
+use anyhow::{Context as _, IntoResult, Result};
+use chrono::NaiveTime;
+use logos::{Lexer, Logos};
 use twilight_http::request::channel::reaction::RequestReactionType;
-use twilight_mention::{
-    timestamp::{Timestamp, TimestampStyle},
-    Mention,
-};
+use twilight_mention::Mention;
 use twilight_model::{
     channel::{ChannelType, Message},
     guild::Permissions,
@@ -14,7 +12,7 @@ use twilight_model::{
 };
 use twilight_webhook::util::{MinimalMember, MinimalWebhook};
 
-use crate::{database, Context};
+use crate::{database, parse::token::Format, Context};
 
 /// the reaction request with the unknown timezone emoji
 const UNKNOWN_TIMEZONE_EMOJI: RequestReactionType = RequestReactionType::Custom {
@@ -23,7 +21,7 @@ const UNKNOWN_TIMEZONE_EMOJI: RequestReactionType = RequestReactionType::Custom 
 };
 
 /// whether the time is am or pm, if 12-hour
-#[derive(Clone, Copy)]
+#[derive(Clone, Copy, Debug, PartialEq)]
 pub enum AmPm {
     /// time is am
     Am,
@@ -31,38 +29,57 @@ pub enum AmPm {
     Pm,
 }
 
-/// whether the time is 12-hour or 24-hour format
-#[derive(Clone, Copy)]
-pub enum Format {
-    /// time is in 12-hour format
-    Hour12,
-    /// time is in 24-hour format
-    Hour24,
-}
+/// in a module to allow lints in the derive macro
+#[allow(clippy::use_self, clippy::indexing_slicing)]
+pub mod token {
+    use chrono::{NaiveTime, Utc};
+    use chrono_tz::Tz;
+    use logos::Logos;
+    use twilight_mention::timestamp::{Timestamp, TimestampStyle};
 
-/// returns the regex to parse times in 12 hour format
-pub fn regex_12_hour() -> Result<Regex> {
-    Ok(RegexBuilder::new(
-        r"(?:^|\s)(?P<hour>1[0-2]|0?[1-9])(?::(?P<minute>[0-5]\d))?\s*(?P<am_pm>am|pm)\b",
-    )
-    .case_insensitive(true)
-    .multi_line(true)
-    .build()?)
-}
+    use super::{hour_12, hour_12_minute, hour_24};
 
-/// returns the regex to parse times in 24 hour format
-pub fn regex_24_hour() -> Result<Regex> {
-    Ok(
-        RegexBuilder::new(r"(?:^|\s)(?P<hour>[0-1]?\d|2[0-3]):(?P<minute>[0-5]\d)\b")
-            .case_insensitive(true)
-            .multi_line(true)
-            .build()?,
-    )
+    /// the format of the time
+    #[derive(Logos, Debug, PartialEq)]
+    pub enum Format {
+        /// a format like `14:33`
+        #[regex(r"\d\d?:\d\d", hour_24)]
+        Hour24(NaiveTime),
+        /// a format like `11am`
+        #[regex(r"\d\d? ?(?:am|pm|Am|Pm|AM|PM)", hour_12)]
+        Hour12(NaiveTime),
+        /// a format like `11:33am`
+        #[regex(r"\d\d?:\d\d ?(?:am|pm|Am|Pm|AM|PM)", hour_12_minute)]
+        Hour12Minute(NaiveTime),
+        /// required by logos
+        #[error]
+        Error,
+    }
+
+    impl Format {
+        /// returns the timestamp of the parsed time, if any
+        pub fn timestamp(self, tz: Tz) -> Option<Timestamp> {
+            if let Format::Hour24(time) | Format::Hour12(time) | Format::Hour12Minute(time) = self {
+                Some(Timestamp::new(
+                    Utc::today()
+                        .with_timezone(&tz)
+                        .and_time(time)?
+                        .timestamp()
+                        .try_into()
+                        .ok()?,
+                    Some(TimestampStyle::ShortTime),
+                ))
+            } else {
+                None
+            }
+        }
+    }
 }
 
 /// adds discord formatted timestamp after times from message content and its
 /// author's saved timezone and sends it to the message's channel impersonating
 /// the author
+#[allow(clippy::integer_arithmetic, unused_must_use)]
 pub async fn send_time(ctx: Context, message: Message) -> Result<()> {
     if message.author.bot
         || ctx
@@ -83,28 +100,6 @@ pub async fn send_time(ctx: Context, message: Message) -> Result<()> {
         return Ok(());
     }
 
-    let mut top_captures: Vec<(Captures, Format)> = ctx
-        .regex_12_hour
-        .captures_iter(&message.content)
-        .map(|captures| (captures, Format::Hour12))
-        .chain(
-            ctx.regex_24_hour
-                .captures_iter(&message.content)
-                .map(|captures| (captures, Format::Hour24)),
-        )
-        .collect();
-
-    top_captures.sort_unstable_by(|old, new| {
-        old.0
-            .get(0)
-            .map_or(0, |m| m.start())
-            .cmp(&new.0.get(0).map_or(0, |m| m.start()))
-    });
-
-    if top_captures.is_empty() {
-        return Ok(());
-    };
-
     let tz = if let Some(tz) = database::timezone(&ctx.db, message.author.id).await? {
         tz
     } else {
@@ -115,17 +110,20 @@ pub async fn send_time(ctx: Context, message: Message) -> Result<()> {
         return Ok(());
     };
 
-    let mut content = String::with_capacity(70);
-    let mut last_pushed_idx = 0;
-    for captures in top_captures {
-        let timestamp = time(&captures.0, captures.1, tz)?;
-
-        let push_idx = captures.0.iter().last().ok()?.ok()?.end();
-        let old_content = message.content.get(last_pushed_idx..push_idx).ok()?;
-        last_pushed_idx = push_idx;
-        content.push_str(&format!("{} ({})", old_content, timestamp.mention()));
+    let lex = Format::lexer(&message.content).spanned();
+    let mut content = String::with_capacity(message.content.len() + 70);
+    let mut pushed_until = 0;
+    for (format, span) in lex {
+        if let Some(timestamp) = format.timestamp(tz) {
+            content.push_str(message.content.get(pushed_until..span.end).ok()?);
+            pushed_until = span.end;
+            write!(content, " ({})", timestamp.mention());
+        };
     }
-    content.push_str(message.content.get(last_pushed_idx..).ok()?);
+    content.push_str(message.content.get(pushed_until..).ok()?);
+    if content.is_empty() || content == message.content {
+        return Ok(());
+    }
 
     ctx.http
         .delete_message(message.channel_id, message.id)
@@ -156,63 +154,6 @@ pub async fn send_time(ctx: Context, message: Message) -> Result<()> {
     Ok(())
 }
 
-/// parses a time from captures
-pub fn time(captures: &Captures, format: Format, tz: Tz) -> Result<Timestamp> {
-    match format {
-        Format::Hour12 => parse_12_hour(captures, tz),
-        Format::Hour24 => parse_24_hour(captures, tz),
-    }
-}
-
-/// parses a time in 24 hour format from captures
-fn parse_24_hour(captures: &Captures, tz: Tz) -> Result<Timestamp> {
-    let hour: u32 = captures.name("hour").ok()?.as_str().parse()?;
-    let minute: u32 = captures
-        .name("minute")
-        .map_or("0", |m| m.as_str())
-        .parse()?;
-
-    Ok(Timestamp::new(
-        Utc::today()
-            .with_timezone(&tz)
-            .and_hms_opt(hour, minute, 0)
-            .ok()?
-            .timestamp()
-            .try_into()?,
-        Some(TimestampStyle::ShortTime),
-    ))
-}
-
-/// parses a time in 12 hour format from captures
-fn parse_12_hour(captures: &Captures, tz: Tz) -> Result<Timestamp> {
-    let hour: u32 = captures.name("hour").ok()?.as_str().parse()?;
-    let minute: u32 = captures
-        .name("minute")
-        .map_or("0", |m| m.as_str())
-        .parse()?;
-    let am_pm = match captures
-        .name("am_pm")
-        .ok()?
-        .as_str()
-        .to_lowercase()
-        .as_ref()
-    {
-        "am" => AmPm::Am,
-        "pm" => AmPm::Pm,
-        _ => bail!("am_pm capture isn't am or pm"),
-    };
-
-    Ok(Timestamp::new(
-        Utc::today()
-            .with_timezone(&tz)
-            .and_hms_opt(to_24_hour(hour, am_pm).ok()?, minute, 0)
-            .ok()?
-            .timestamp()
-            .try_into()?,
-        Some(TimestampStyle::ShortTime),
-    ))
-}
-
 /// converts 12-hour to 24-hour format
 #[allow(clippy::integer_arithmetic)]
 pub const fn to_24_hour(hour: u32, am_pm: AmPm) -> Option<u32> {
@@ -236,4 +177,38 @@ pub const fn to_24_hour(hour: u32, am_pm: AmPm) -> Option<u32> {
             }
         }
     })
+}
+
+/// parses a time in 24 hour format
+fn hour_24(lex: &mut Lexer<Format>) -> Option<NaiveTime> {
+    let (hour, minute) = lex.slice().split_once(':')?;
+    NaiveTime::from_hms_opt(hour.parse().ok()?, minute.parse().ok()?, 0)
+}
+
+/// parses a time in 12 hour format
+fn hour_12(lex: &mut Lexer<Format>) -> Option<NaiveTime> {
+    let slice = lex.slice();
+    let (hour, _) = slice.split_once(&['a', 'A', 'p', 'P'])?;
+    let am_pm = if slice.ends_with("am") || slice.ends_with("Am") || slice.ends_with("AM") {
+        AmPm::Am
+    } else {
+        AmPm::Pm
+    };
+    NaiveTime::from_hms_opt(to_24_hour(hour.trim().parse().ok()?, am_pm)?, 0, 0)
+}
+
+/// parses a time with minute in 12 hour format
+fn hour_12_minute(lex: &mut Lexer<Format>) -> Option<NaiveTime> {
+    let (hour, after) = lex.slice().split_once(':')?;
+    let minute = after.get(..2)?;
+    let am_pm = if after.ends_with("am") || after.ends_with("Am") || after.ends_with("AM") {
+        AmPm::Am
+    } else {
+        AmPm::Pm
+    };
+    NaiveTime::from_hms_opt(
+        to_24_hour(hour.parse().ok()?, am_pm)?,
+        minute.parse().ok()?,
+        0,
+    )
 }
